+++ finder/nodejs/lib/base64url.ts
@@ -1,4 +1,5 @@
 const W3C_ELEMENT: string = `element-6066-11e4-a52e-4f735466cecf`;
+const MJSON_ELEMENT: string = `ELEMENT`;
 
 export const encode = (input: string) =>
   Buffer.from(input)
@@ -7,16 +8,14 @@ export const encode = (input: string) =>
     .replace(/\+/g, `-`)
     .replace(/\//g, `_`);
 
-export const decode = (input: string | {ELEMENT: string}) => {
+export const decode = (input: string | {[key: string]: string}) => {
   let base64String: string = ``;
   if (typeof input === `string`) {
     base64String = input;
-  // @ts-ignore
-  } else if (typeof input === `object` && input[W3C_ELEMENT]) {
-    // @ts-ignore
+  } else if (typeof input === `object` && W3C_ELEMENT in input) {
     base64String = input[W3C_ELEMENT];
-  } else if (typeof input === `object` && input.ELEMENT) {
-    base64String = input.ELEMENT;
+  } else if (typeof input === `object` && MJSON_ELEMENT in input) {
+    base64String = input[MJSON_ELEMENT];
   } else {
     throw new Error(`input is invalid ${JSON.stringify(input)}`);
   }